requirements.txt
backcall==0.1.0
beautifulsoup4==4.8.2
braintree==3.57.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
django-taggit==0.22.2
Flask==1.0.3
idna==2.8
ipython==7.6.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
lxml==4.4.2
Markdown==2.6.11
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
oauthlib==3.0.2
parso==0.5.0
pickleshare==0.7.5
Pillow==5.1.0
pipenv==2018.10.13
prompt-toolkit==2.0.9
psycopg2==2.7.4
Pygments==2.4.2
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==2.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.4.1
soupsieve==1.9.5
traitlets==4.3.2
urllib3==1.25.3
virtualenv==16.6.1
virtualenv-clone==0.5.3
wcwidth==0.1.7
Werkzeug==0.15.4
